build: add write permission for pages
I added the "pages" permission to the GitHub workflow file `.github/workflows/release-please.yml`. This allows the workflow to write to repository pages. To authenticate, a personal access token (PAT) named `MY_RELEASE_PLEASE_TOKEN` needs to be configured as a secret in the GitHub actions settings. I have already done this.